refactor: move download receipt code to openEappcontroller
